EXC-1225: Update code according to spec http requests

This MR adds makes some changes in the replica to be consistent with the IC spec for http_request:

* `status: nat` with the corresponding rust type:  u128 
* method : variant { get; head; post };  use lowercase and use "method" (all as in spec)
* transform function: Use `candid:Func` which provides PrincipalId too; and check that the principal is equivalent to the sender id. 

See merge request dfinity-lab/public/ic!7284

+++ rs/rust_canisters/proxy_canister/src/lib.rs
@@ -16,13 +16,13 @@ pub struct RemoteHttpRequest {
 
 #[derive(CandidType, Deserialize, Debug, Clone)]
 pub struct RemoteHttpResponse {
-    pub status: u8,
+    pub status: u128,
     pub headers: Vec<(String, String)>,
     pub body: String,
 }
 
 impl RemoteHttpResponse {
-    pub fn new(status: u8, headers: Vec<(String, String)>, body: String) -> Self {
+    pub fn new(status: u128, headers: Vec<(String, String)>, body: String) -> Self {
         Self {
             status,
             headers,
